Create separate package for calculator + use KeyboardListener instead of RawKeyboardListener
