+++ .gitignore
@@ -21,3 +21,4 @@ lib/generated_plugin_registrant.dart
 debug.log
 .snap
 packages/translations/lib/app_localizations*.dart
+pubspec_overrides.yaml
