+++ melos.yaml
@@ -10,7 +10,7 @@ command:
     # https://github.com/dart-lang/pub/issues/3404). Disabling this feature
     # makes the CI much more stable.
     runPubGetInParallel: false
-  
+    usePubspecOverrides: true
 
 scripts:
   postbootstrap: cd packages/translations && flutter gen-l10n
