+++ packages/calculator_widget/analysis_options.yaml
@@ -1,0 +1,4 @@
+include: package:flutter_lints/flutter.yaml
+
+linter:
+  rules:
