+++ packages/calculator_widget/pubspec.yaml
@@ -1,0 +1,29 @@
+name: calculator_widget
+description: Calculator widget for Converter NOW
+
+publish_to: 'none'
+
+version: 1.0.0+1
+
+environment:
+  sdk: ">=2.17.6 <3.0.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+  #Needed for calculator computation
+  decimal: ^2.2.0
+  #Needed for the state management
+  provider: ^6.0.2
+  #Needed for app translation
+  translations:
+    path: ../translations
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  test: ^1.17.12
+  flutter_lints: ^2.0.0
+
+flutter:
+  uses-material-design: true
